Move SwiftASTContext & TypeSystemSwift* into Plugins/TypeSystem/Swift
to mirror a similar move spearheaded by the Clang typesystem.

(cherry picked from commit 9161b3cb453064096ccb92d181d996154d24ada7)

 Conflicts:
	lldb/source/API/SystemInitializerFull.cpp
	lldb/source/Core/Module.cpp
	lldb/source/Plugins/ExpressionParser/Clang/ClangPersistentVariables.cpp
	lldb/source/Plugins/InstrumentationRuntime/MainThreadChecker/InstrumentationRuntimeMainThreadChecker.cpp
	lldb/source/Plugins/Language/Swift/SwiftDictionary.cpp
	lldb/source/Plugins/Language/Swift/SwiftSet.cpp
	lldb/source/Plugins/TypeSystem/Swift/SwiftASTContext.h
	lldb/source/Target/Target.cpp
	lldb/unittests/Symbol/CMakeLists.txt

(cherry picked from commit a794ee173962e49ac2b9a0b5e8869b6611c21584)
