+++ lldb/source/Plugins/ExpressionParser/Swift/SwiftASTManipulator.cpp
@@ -12,9 +12,9 @@
 
 #include "SwiftASTManipulator.h"
 
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
 #include "lldb/Expression/ExpressionParser.h"
 #include "lldb/Expression/ExpressionSourceCode.h"
-#include "lldb/Symbol/SwiftASTContext.h"
 #include "lldb/Target/Target.h"
 #include "lldb/Utility/ConstString.h"
 #include "lldb/Utility/Log.h"
