+++ lldb/source/Plugins/Language/Swift/SwiftDictionary.cpp
@@ -13,8 +13,8 @@
 #include "SwiftDictionary.h"
 
 #include "Plugins/TypeSystem/Clang/TypeSystemClang.h"
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
 #include "lldb/DataFormatters/FormattersHelpers.h"
-#include "lldb/Symbol/SwiftASTContext.h"
 #include "lldb/Target/Process.h"
 #include "lldb/Target/SwiftLanguageRuntime.h"
