+++ lldb/source/Plugins/Language/Swift/SwiftSet.cpp
@@ -13,8 +13,8 @@
 #include "SwiftSet.h"
 
 #include "Plugins/TypeSystem/Clang/TypeSystemClang.h"
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
 #include "lldb/DataFormatters/FormattersHelpers.h"
-#include "lldb/Symbol/SwiftASTContext.h"
 #include "lldb/Target/Process.h"
 #include "lldb/Target/SwiftLanguageRuntime.h"
