+++ lldb/source/Plugins/Language/Swift/SwiftUnsafeTypes.cpp
@@ -1,7 +1,7 @@
 #include "SwiftUnsafeTypes.h"
 
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
 #include "lldb/DataFormatters/TypeSynthetic.h"
-#include "lldb/Symbol/SwiftASTContext.h"
 #include "lldb/Target/SwiftLanguageRuntime.h"
 #include "lldb/Utility/Log.h"
 #include "lldb/Utility/Logging.h"
