+++ lldb/source/Plugins/TypeSystem/CMakeLists.txt
@@ -1,1 +1,5 @@
 add_subdirectory(Clang)
+
+if (LLDB_ENABLE_SWIFT_SUPPORT)
+  add_subdirectory(Swift)
+endif()
