+++ lldb/source/Plugins/TypeSystem/Swift/CMakeLists.txt
@@ -1,0 +1,14 @@
+add_lldb_library(lldbPluginTypeSystemSwift PLUGIN
+  TypeSystemSwift.cpp
+  TypeSystemSwiftTypeRef.cpp
+  SwiftASTContext.cpp
+
+  LINK_LIBS
+    lldbCore
+    lldbSymbol
+    lldbTarget
+    lldbUtility
+
+  LINK_COMPONENTS
+    Support
+)
