+++ lldb/unittests/Symbol/CMakeLists.txt
@@ -27,6 +27,7 @@ add_lldb_unittest(SymbolTests
     lldbPluginSymbolFileDWARF
     lldbPluginSymbolFileSymtab
     lldbPluginTypeSystemClang
+    lldbPluginTypeSystemSwift
     LLVMTestingSupport
   )
