+++ lldb/unittests/Symbol/TestSwiftASTContext.cpp
@@ -14,7 +14,7 @@
 
 #include "lldb/Host/FileSystem.h"
 #include "lldb/Host/HostInfo.h"
-#include "lldb/Symbol/SwiftASTContext.h"
+#include "Plugins/TypeSystem/Swift/SwiftASTContext.h"
 
 using namespace lldb;
 using namespace lldb_private;
