+++ lldb/unittests/Symbol/TestTypeSystemSwiftTypeRef.cpp
@@ -12,7 +12,8 @@
 
 #include "gtest/gtest.h"
 
-#include "lldb/Symbol/SwiftASTContext.h"
+#include "Plugins/TypeSystem/Swift/TypeSystemSwiftTypeRef.h"
+#include "llvm/ADT/StringRef.h"
 #include "swift/Demangling/Demangle.h"
 #include "swift/Demangling/Demangler.h"
 #include "swift/Strings.h"
